No implementation.hpp
